Move wav Config to separate file
